+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0
+**10.09.2024**
+
+- Widevine device new parse variant support: device_client_id_blob and device private key (PEM)
+- Widevine device info (device.wvd file) serializer
+
 ## 1.0.3
 **27.08.2024**